requirements.txt
fastapi
uvicorn[standard]
aiohttp
numpy
xarray
cfgrib
pillow
mercantile
rasterio
matplotlib
